requirements.txt
streamlit>=1.33.0
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.31.0
tqdm>=4.66.0